notfinal which do not make polygon added.

+++ src/o5m2sqlite.c
@@ -540,20 +540,25 @@ int main(int argc, char **argv) {
 		
 	}				
 
-	sqlite3_exec(db,"DROP TABLE polygon",0,0,0);
+	if ( argc > 2 && argv[2] == 'notfinal' ) {
+	}
+	else {
 
-	sqlite3_exec(db,
-		"CREATE TABLE polygon AS "
-		"SELECT id,geom,0 AS is_rel FROM way WHERE closed = 1 AND IsValid(geom) AND NOT IsEmpty(geom) AND NumPoints(geom) > 3 "
-		"UNION "
-		"SELECT * FROM ( "
-		"SELECT rel_id,Collect(geom) AS geom,1 AS is_rel FROM "
-		"(SELECT rel_id,LineMerge(Collect(geom)) AS geom,object FROM rel_outer RO "
-		"JOIN way W ON W.id = RO.way_id AND IsValid(W.geom) AND NOT IsEmpty(W.geom) AND NumPoints(W.geom) > 1 "		
-		"GROUP BY rel_id, object "
-		"ORDER BY segment) "
-		"GROUP BY rel_id)"	
-	,0,0,0);
+		sqlite3_exec(db,"DROP TABLE polygon",0,0,0);
+
+		sqlite3_exec(db,
+			"CREATE TABLE polygon AS "
+			"SELECT id,geom,0 AS is_rel FROM way WHERE closed = 1 AND IsValid(geom) AND NOT IsEmpty(geom) AND NumPoints(geom) > 3 "
+			"UNION "
+			"SELECT * FROM ( "
+			"SELECT rel_id,Collect(geom) AS geom,1 AS is_rel FROM "
+			"(SELECT rel_id,LineMerge(Collect(geom)) AS geom,object FROM rel_outer RO "
+			"JOIN way W ON W.id = RO.way_id AND IsValid(W.geom) AND NOT IsEmpty(W.geom) AND NumPoints(W.geom) > 1 "		
+			"GROUP BY rel_id, object "
+			"ORDER BY segment) "
+			"GROUP BY rel_id)"	
+		,0,0,0);
+	}
 
 	sqlite3_exec(db,"COMMIT TRANSACTION",0,0,0);	
 	sqlite3_exec(db, "VACUUM",0,0,0);
